refactor: shorten config command file names
